vscode-shell: extract bridge helpers into src/bridge/
Split AgentServerBridge by moving low-coupled chunks out:

  - messages.ts: BridgeTo/FromWebviewMessage type unions

  - shim.ts: agent-server *Session->*Conversation compat layer

  - requestIds.ts: clientIdOf helper

  - clientIO.ts: createBridgeClientIO factory with narrow context

    (broadcast / rememberServerRequestId / handleShellAction)

  - historyReplay.ts: pure toHistoryReplayMessage mapping

agentServerBridge.ts: 1818 -> 1430 LOC (no behavior change).

Connection lifecycle, session ops, and replay-buffer state stay in

the bridge - those flows are tightly coupled and ordering-sensitive.
